+++ src/app/(auth)/layout.tsx
@@ -1,5 +1,5 @@
 import { AuthProvider } from '@/context/auth-context';
-import { AuthGuard } from '@/components/auth-guard/auth-guard';
+import { AuthGuard } from '@/guards/auth-guard';
 import React, { Suspense } from 'react';
 
 export default function AuthLayout({ children }: { children: React.ReactNode }) {